requirements.txt
pytest
numpy
pandas
matplotlib
requests
biopython
snakemake
